+++ src/command_modules/azure-cli-cosmosdb/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+
+unreleased
++++++++++++++++++++
+* Minor fix allowing Creation of Collection with custom partition key
+
 0.1.10 (2017-07-07)
 +++++++++++++++++++
 * minor fixes